+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - [9046](https://github.com/vegaprotocol/vega/issues/9046) - Send event on referral engine state change.
 - [9045](https://github.com/vegaprotocol/vega/issues/9045) - Snapshot the referral engine.
 - [9136](https://github.com/vegaprotocol/vega/issues/9136) - Referral engine returns reward factor for a given party
+- [9076](https://github.com/vegaprotocol/vega/issues/9076) - Add current funding rate to market data.
 
 ### 🐛 Fixes